apps/consumers/vital-hl7/requirements.txt
kafka-python>=2.0.2
pymongo>=4.4.0
lz4>=4.3.2